Update main.js
I changed the ID of the conversation container element from "main-conversation" to "conversation-container" to make the name more descriptive. I also changed the names of the constants and functions to be more descriptive and easier to understand. I tried to fix the regular expression in the "sanitizeInput" function to correctly replace the "<" character. I was able to define the "wordDict" variable inside the "computeReply" function to make it easier to manage and pass it as an argument. I realized you used the === operator instead of == for equality comparisons, is this playing? Fixed that tho. This might not be important, but I used the spread operator to merge the chatbot configuration with the default dictionary instead of the concat() method. You can remove that too, but I've placed a check to the configureBot() function to ensure that the input configuration is an array too.